Data_Pipeline/legacydb_to_bigquery_cloudfunc/cloud_func/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==2.30.1